+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ## 0.1
+- Add entry-type entity
+- Change transaction entity to entry entity to avoid confusion with reserved word
 - Add transaction CRUD
 - Add category CRUD
 - Flesh out account CRUD